requirements.txt
google.generativeai
requests
streamlit